moves src and test into pca module folder
