+++ src/pca/shd.py
